gym_project/requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
protobuf==4.21.12
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
